+++ logger.py
@@ -1,6 +1,4 @@
-import logging
-import os
-import inspect
+import logging, os, inspect
 from config import Config
 from logging.handlers import RotatingFileHandler
 
@@ -22,22 +20,26 @@ logger = logging.getLogger()
 logger.setLevel(logging.INFO)
 logger.addHandler(handler)
 
-def log_it(level, msg):
+def log_it(level, msg, exc=None):
     caller_frame = inspect.currentframe().f_back
     func_name = caller_frame.f_code.co_name
+    filename = caller_frame.f_code.co_filename
 
     if level.lower() == "d" and not Config.DEBUG:
         return
+
     levels = {
         "i": logging.info,
         "e": logging.error,
-        "d": logging.info, # Use info because debug flag spams so much
+        "d": logging.info,
     }
 
     log_func = levels.get(level.lower(), None)
 
     if log_func:
-        if level.lower() == "d":
+        if exc:
+            log_func(f"[{func_name}] [{filename}] {msg} - Exception: {exc}")
+        elif level.lower() == "d":
             log_func(f"[DEBUG] [{func_name}] {msg}")
         else:
             log_func(f"[{func_name}] {msg}")